Added timer to bombs and missiles player1

+++ SDL_AndroDunos/ModuleParticles.cpp
@@ -733,9 +733,9 @@ bool Particle::Update()
 		if (target == nullptr)
 			target = FindE(position);
 		else {
-			int x = target->position.x - position.x;
-			int y = target->position.y - position.y;
-			int modulev = sqrt(x*x + y*y);
+			float x = target->position.x - position.x;
+			float y = target->position.y - position.y;
+			float modulev = sqrt(x*x + y*y);
 			x /= modulev;
 			y /= modulev;
 			position.x += x * 5;
